requirements.txt
# -------------------------------
# requirements.txt – LADP chatbot
# -------------------------------

# --- UI & core -------------------------------------------------------------
streamlit>=1.34
python-dotenv

# --- Data wrangling --------------------------------------------------------
pandas>=2.0
numpy>=1.24          # optional – keeps pip-deptree happy

# --- LLM stack (Azure & public OpenAI) ------------------------------------
openai>=1.25,<2       # 1.59.6 tested, keep <2 ceiling for breaking changes
langchain==0.3.4      # pin until next refactor
langchain-openai>=0.2.3

# --- Vector store (pick ONE) ----------------------------------------------
faiss-cpu             # comment-out if you migrate to Chroma below
# chromadb
# langchain-chroma>=0.0.5

# --- PDF parsing -----------------------------------------------------------
PyPDF2==3.0.1         # swap to pypdf==3.17.0 if preferred but remove PyPDF2
pdfplumber>=0.11
# pypdf==3.17.0

# --- Tokeniser -------------------------------------------------------------
tiktoken>=0.6.0 

# --- Visualisation -------------------------------------------------
matplotlib>=3.8          # for pie-chart rendering

# ---- Yahoo Finance -------------------------------------------------
yfinance>=0.2.6